requirements.txt
python-dotenv==1.0.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bidict==0.23.1
bitarray==3.5.2
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
cytoolz==1.0.1
dnspython==2.7.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
eventlet==0.40.1
fastapi==0.116.1
Flask==3.1.1
flask-cors==6.0.1
Flask-SocketIO==5.5.1
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
hexbytes==1.3.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
multidict==6.6.3
parsimonious==0.10.0
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-engineio==4.12.2
python-socketio==5.13.0
pyunormalize==16.0.0
pywin32==311
regex==2024.11.6
requests==2.32.4
rlp==4.1.0
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.47.2
toolz==1.0.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.22.0
web3==7.12.1
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
git+https://github.com/TTD-dev/TTD-dev.git
